Fix pagination when some pages are hidden
The logic for determining the successor of a page, used for determining
the links to the previous and next pages, would return hidden pages,
which was not intentional. This is fixed by skipping hidden pages in all
situations unconditionally.

Also, hidden pages would display a pagination bar, which doesn’t really
make sense since hidden pages aren’t part of the menu and, hence, don’t
have predecessor and successor pages. For hidden pages, the pagination
bar is now hidden altogether.