Improve layout of display elements
Because of how the CSS line-height property works (it increases the
vertical space consumed by each line by adding space above and below
each line), display elements (such as figures, tables, and code blocks)
would appear to have smaller margins above and below them than regular
text. This is because display elements aren’t assigned the additional
space that regular body text gets as a result of the line-height
property.

To fix this, this commit introduces a margin correction function. Added
to the regular margins around display elements, the vertical spaces
between blocks in the main column appear more even. The only known
exception occurs when there are two consecutive display elements. In
such situations, an additional margin correction would be needed.
However, the current spectrum of CSS properties doesn’t make it easy to
fix that, so this limitation has to be accepted for now.

Additionally, this removes margins introduced by the first and last
children of the main column, blockquotes, and lists.