Use anchorize function for section anchors
The section anchors inserted by the alert shortcode used the urlize
function to convert the title into a normalized form (using lowercase
and dashes). However, urlize is intended to be used for generating URLs,
and there is a specialized function anchorize for generating section
anchor names, so use that one instead.