Support font icons in navigation sidebar
This allows users to specify an icon to be shown next to menu bar items.
The icon is configured using the custom icon site variable.

Configuring icons with image URLs is no longer supported. If users wish
to include custom icons not available in Tabler Icons, it’s recommended
to create a new icon font to that end.