Make page title optional
By providing the new “hideTitle” page parameter, the title of a page can
be hidden. This is useful in combination with the new “featuredImage”
page parameter in the event that an author wants to show a title banner
in place of the regular title. (Otherwise, the title would be shown
redundantly in the featured image and the regular page title.)